+++ frontend/src/solid-ui/index.ts
@@ -30,3 +30,4 @@ export * from "./MarqueeText";
 export * from "./DitheredImage";
 export * from "./meshTint";
 export * from "./AppLogo";
+export * from "./LoadingState";